Map functionality complete pending vendor updates/merging
Current functionality: Clicking on Vendors, then on Map brings up the map page. Pins are dynamically generated based on the addresses of all of the vendors that our geocoding service can find lat/long for (shitty addresses = no map pin). Clicking on the pin brings up the vendor name, address, and a link to the vendor page. Clicking on the link to the vendor page redirects to my_vendors/[vendor_id]. Also, our API key is secured and encrypted (once I delete the branch where I accidentally pushed it unsecured).

Outstanding bugs / features left to be added: Following the link to the vendor page gives an error related to the ownership variable instead of displaying the page. Pretty sure this should be fixed once your vendors stuff is added in. I also still need to add a link from the address on the vendor page to the map, but I didn't want to do so until I had access to the finalized vendor pages.